+++ next/env.ts
@@ -4,7 +4,6 @@ import zod from "zod";
 export const env = createEnv({
   server: {
     NODE_ENV: zod.enum(["development", "production", "test"]),
-    DRUPAL_SITE_ID: zod.string().optional(),
     DRUPAL_CLIENT_ID: zod.string(),
     DRUPAL_CLIENT_SECRET: zod.string(),
     DRUPAL_CLIENT_VIEWER_ID: zod.string(),
@@ -20,7 +19,6 @@ export const env = createEnv({
   /* eslint-disable n/no-process-env */
   runtimeEnv: {
     NODE_ENV: process.env.NODE_ENV,
-    DRUPAL_SITE_ID: process.env.DRUPAL_SITE_ID,
     DRUPAL_CLIENT_ID: process.env.DRUPAL_CLIENT_ID,
     DRUPAL_CLIENT_SECRET: process.env.DRUPAL_CLIENT_SECRET,
     DRUPAL_CLIENT_VIEWER_ID: process.env.DRUPAL_CLIENT_VIEWER_ID,