+++ next/lib/get-common-page-props.ts
@@ -3,12 +3,18 @@ import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 
 import { getMenus } from "@/lib/drupal/get-menus";
 
+import siteConfig from "@/site.config";
+
 export type CommonPageProps = Awaited<ReturnType<typeof getCommonPageProps>>;
 
-export async function getCommonPageProps(context: GetStaticPropsContext) {
+export async function getCommonPageProps({
+  locale = siteConfig.defaultLocale,
+}: {
+  locale: GetStaticPropsContext["locale"];
+}) {
   const [translations, menus] = await Promise.all([
-    serverSideTranslations(context.locale),
-    getMenus(context),
+    serverSideTranslations(locale),
+    getMenus({ locale }),
   ]);
 
   return {