+++ next/next-i18next.config.js
@@ -5,4 +5,13 @@ module.exports = {
     defaultLocale: config.defaultLocale,
     locales: Object.keys(config.locales),
   },
+  ns: "common",
+  defaultNS: "common",
+  fallbackNS: "common",
+  react: {
+    useSuspense: false,
+  },
+  interpolation: {
+    escapeValue: false,
+  },
 };