+++ next/pages/500.tsx
@@ -4,6 +4,7 @@ import { useTranslation } from "next-i18next";
 
 import { HeadingPage } from "@/components/heading--page";
 import { Meta } from "@/components/meta";
+import { REVALIDATE_LONG } from "@/lib/constants";
 import {
   CommonPageProps,
   getCommonPageProps,
@@ -25,13 +26,13 @@ export default function NotFoundPage() {
   );
 }
 
-export const getStaticProps: GetStaticProps<CommonPageProps> = async (
-  context,
-) => {
+export const getStaticProps: GetStaticProps<CommonPageProps> = async ({
+  locale,
+}) => {
   return {
     props: {
-      ...(await getCommonPageProps(context)),
+      ...(await getCommonPageProps({ locale })),
     },
-    revalidate: 60,
+    revalidate: REVALIDATE_LONG,
   };
 };